requirements.txt
absl-py @ file:///C:/ci/absl-py_1615411407614/work
aiohttp @ file:///C:/ci/aiohttp_1602530358780/work
altair==4.2.0
altgraph==0.17
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blinker==1.4
bokeh==2.4.3
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet==4.0.0
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
click-plugins==1.1.1
cligj==0.7.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
commonmark==0.9.1
coverage @ file:///C:/ci/coverage_1614614910274/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
cycler==0.10.0
Cython @ file:///C:/ci/cython_1614015037049/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distro==1.5.0
dnspython==2.2.1
entrypoints==0.3
et-xmlfile==1.0.1
Fiona==1.8.13.post1
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
GDAL==3.0.2
geographiclib==1.52
geopandas @ file:///tmp/build/80754af9/geopandas_1594937497867/work
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.29
google-auth @ file:///tmp/build/80754af9/google-auth_1616008050444/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1614884419385/work
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1615900693889/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1614616717881/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib==1.0.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.1
Levenshtein==0.20.5
livelossplot==0.5.5
lxml==4.6.2
Markdown @ file:///C:/ci/markdown_1614364005059/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib==3.3.4
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1600456486794/work
munch==2.5.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1611348264852/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.0
opencv-python==4.6.0.66
openpyxl==3.0.7
opt-einsum==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1614711338358/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pdf2image==1.14.0
pdfminer.six==20201018
pdftotext==2.1.5
pefile==2019.4.18
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
psycopg2==2.9.3
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.8.0
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
PyGreSQL==5.2.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pymongo==4.2.0
Pympler==1.0.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PyPDF2==1.26.0
pyproj @ file:///C:/ci/pyproj_1614278152133/work
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytesseract==0.3.7
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-Levenshtein==0.20.5
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
rapidfuzz==2.11.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rich==12.6.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
Rtree==0.9.4
scikit-learn==0.24.1
scipy @ file:///C:/ci/scipy_1616703425152/work
seaborn==0.11.2
semver==2.13.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
Shapely==1.7.1
six @ file:///C:/ci/six_1605205426665/work
smmap==5.0.0
sortedcontainers==2.3.0
streamlit==1.14.0
tabula-py==2.2.0
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.9.3
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
thefuzz==0.19.0
threadpoolctl==2.1.0
tika==1.24
toml==0.10.2
toolz==0.12.0
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==4.3.0
tzdata==2022.6
tzlocal==4.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
validators==0.20.0
watchdog==2.1.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.3
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
